requirements.txt
APScheduler==3.10.4
python-dateutil==2.8.2
SQLAlchemy==2.0.25
requests==2.31.0
python-dotenv==1.0.1
psutil==5.9.8
PyJWT==2.8.0
bcrypt==4.1.2
fastapi==0.109.0
uvicorn==0.27.0
pydantic[email]==2.6.1
email-validator==2.1.0.post1